requirements.txt
numpy==1.26.2
pandas==2.1.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.23